requirements.txt
requests
kafka-python
google-cloud-bigquery
python-dotenv